share prom logic
