+++ crates/sui-aws-orchestrator/Cargo.toml
@@ -27,6 +27,7 @@ serde.workspace = true
 tokio = { workspace = true, features = ["full"] }
 prometheus-parse.workspace = true
 
+mysten-metrics.workspace = true
 sui-config = { path = "../sui-config" }
 sui-types = { path = "../sui-types" }
 sui-node = { path = "../sui-node" }
