+++ crates/sui-aws-orchestrator/src/protocol/sui.rs
@@ -265,7 +265,7 @@ impl ProtocolMetrics for SuiProtocol {
                     "{}:{}{}",
                     instance.main_ip,
                     config.metrics_address.port(),
-                    sui_node::metrics::METRICS_ROUTE
+                    mysten_metrics::METRICS_ROUTE
                 );
                 (instance, path)
             })
@@ -283,7 +283,7 @@ impl ProtocolMetrics for SuiProtocol {
                     "{}:{}{}",
                     instance.main_ip,
                     Self::CLIENT_METRICS_PORT,
-                    sui_node::metrics::METRICS_ROUTE
+                    mysten_metrics::METRICS_ROUTE
                 );
                 (instance, path)
             })
