+++ crates/sui-benchmark/src/bin/stress.rs
@@ -18,8 +18,6 @@ use sui_benchmark::drivers::BenchmarkCmp;
 use sui_benchmark::drivers::BenchmarkStats;
 use sui_protocol_config::{ProtocolConfig, ProtocolVersion};
 
-use sui_node::metrics;
-
 use sui_benchmark::benchmark_setup::Env;
 use sui_benchmark::options::Opts;
 
@@ -81,7 +79,7 @@ async fn main() -> Result<()> {
     }
     let _guard = config.with_env().init();
 
-    let registry_service = metrics::start_prometheus_server(
+    let registry_service = mysten_metrics::start_prometheus_server(
         format!("{}:{}", opts.client_metric_host, opts.client_metric_port)
             .parse()
             .unwrap(),
