+++ crates/sui-faucet/src/main.rs
@@ -68,7 +68,7 @@ async fn main() -> Result<(), anyhow::Error> {
 
     let prom_binding = PROM_PORT_ADDR.parse().unwrap();
     info!("Starting Prometheus HTTP endpoint at {}", prom_binding);
-    let registry_service = sui_node::metrics::start_prometheus_server(prom_binding);
+    let registry_service = mysten_metrics::start_prometheus_server(prom_binding);
     let prometheus_registry = registry_service.default_registry();
     let app_state = Arc::new(AppState {
         faucet: SimpleFaucet::new(
